debug_long_repr: Cut off long bytes
In practice, when there are large files in the mocks3, printing the `repr` can take quite some time, on the order of 10s!

Instead, only extract the bytes we care, and calculate the repr of those, and then cut that off if necessary, which makes the whole task finish in <100ms.