Fix XSS vulnerability in assert_html_snapshot
We passed `formatter=None` to beautiful_soup.prettify in an effort to not mess with the input.
But this _MASSIVELY_ messes with the input and does not correctly escape it.
This could even lead to Cross-Site-Scripting (XSS) vulnerabilities! Fortunately this function is not used on user input, but still.

Add a test to ensure correct encoding.