Make RAISE_SNAPSHOT_ERRORS overwriteable
It should be possible to overwrite the current `RAISE_SNAPSHOT_ERRORS` environment value. So that a
test can set this value.

This is needed if a test should cover both behaviors and used `OverrideEnviron()` context manager.

Also update tox config, see:
https://github.com/boxine/bx_py_utils/actions/runs/3806591898/jobs/6475523621

And remove this warning:
```
UserWarning: `sections` setting includes EXTERNAL, but no known_external is defined. The following
known_SECTION config options are defined: .
```
because we have to remove the obsolete `.isort.cfg` (The config was added into `pyproject.toml` in
the past)