NEW: Helper to add doctests into unittest by just adding a TestCase
The "normal" way to add doctests to unittests is described here:

https://docs.python.org/3/library/doctest.html#unittest-api

But it's needed to change the unittest setup and how tests loaded. With the `BaseDocTests` it's just
needed to ass another `TestCase` and you are done ;)

This will increase the number of tests from 88 to 167, because we have already many doctests and
after replace pytests with unittests they was not tested. But hey, no changes needed, all doctests
are still correct ;)