Fix NoLogs
NoLogs installed no handlers. But what happens if you don't install any handlers is that Python's logging framework automatically logs into a last-resort handler to stderr.
Fix this by installing a `NullHandler`, and add both a test capturing `sys.stderr` as well as just executing it so we can see it (in case the sys.stderr` capture fails).