Expand DocWrite with optional `delete_obsolete_files`
If DocWrite strings will be remove in code base, so that a *.md file is completely obsolete, then
this file will no be updated and not be removed, because it's just not "seen" and touched.

Add a `delete_obsolete_files` to config to remove all "unseen" *.md files from base path to cleanup.