Switch from pipenv to uv
Replace "pipenv" with "uv" in a minimal way: Keep the existing Makefile targets. So it stays with
the concept to be self contained. This means: You didn't need to have `uv` installed:

Just run `make install` to create a `.venv` and install all needed packages.
The minimal requirements are only `python3-venv` and `python3-pip` (`uv` will be installed via `pip`
in `.venv`)

Also replace "Safety" with https://github.com/pypa/pip-audit because Safety will be not useable in
the future without registration.