remove remainders of pipenv
in 50f34991 pipenv was removed from the project in favor of uv, but `pipenv check` is still used in the publishing script. it effectively does nothing because there is no Pipfile containing dependencies that could be checked.
unfortunately uv doesn't have equivalent functionality (yet), but there are other ways to ensure safe/up-to-date dependencies (e.g. using dependabot).